+++ src/Components/Reviewer/Reviewer.tsx
@@ -1,12 +1,9 @@
 import { FunctionComponent } from 'react';
 import { Dispatch } from 'redux';
 import { useSelector, useDispatch } from 'react-redux';
-import {
-  ErrorAction,
-  loadReviewer,
-  ReviewerAction,
-  SettingsAction,
-} from '../../model/model';
+import { loadReviewer, ReviewerAction } from '../../model/reviewer';
+import { ErrorAction } from '../../model/error';
+import { SettingsAction } from '../../model/settings';
 import { RootState } from '../../store';
 
 import './Reviewer.less';
@@ -28,7 +25,8 @@ const Reviewer: FunctionComponent = () => {
     <div className="content">
       <button
         className="button content__button"
-        disabled={!user || !repo || loading === 'loading'}
+        disabled={!user || !repo || loading}
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         onClick={() => dispatch(loadReviewer())}
       >
@@ -37,28 +35,24 @@ const Reviewer: FunctionComponent = () => {
           : 'Fill user and repo fileds in settings'}
       </button>
 
-      {loading === 'loading' && <div className="spinner"></div>}
+      {loading && <div className="spinner"></div>}
 
-      {loading === 'loaded' && (
+      {data && (
         <div className="content__contributor">
-          {reviewer.data.login ? (
-            <>
-              Your reviewer
-              <a
-                href={data.html_url}
-                className="content__contributor-link"
-                target="_blank"
-              >
-                {data.login}
-              </a>
-            </>
-          ) : (
-            <>There is no contributors for this user or repository</>
-          )}
+          <>
+            Your reviewer
+            <a
+              href={data.html_url}
+              className="content__contributor-link"
+              target="_blank"
+            >
+              {data.login}
+            </a>
+          </>
         </div>
       )}
 
-      {message && code && loading === 'error' && (
+      {code && (
         <div className="content__error">
           Oops! {message} Error status is {code}
         </div>
